Implemented a Simple Integer Index Strategy
Created a simple 1-dimensional NumericIndexStrategy that
can index unsigned longs from 0 to 2^64. The strategy doesn't
use any binning. The ids are simply the byte[] values of
the longs. This index strategy won't work well for inserting
ranges since the data will be replicated for each integer value
in the range.

getQueryRanges always returns 1 range since the sort order of the
bytes is the same as the sort order of the values. This means that
a contiguous range of values can be represented by a single
contiguous ByteArrayRange.

getInsertionIds returns a single insertionId if you are inserting
a point. If you insert a range, you'll get back a list with the
size of the number of integers in the range (because there is no
binning). That is not the intended use of this index strategy.

Included a test for the lexicographical sorting of the byte arrays.
Can't support negative values since the lexicographical ordering
of the byte arrays wouldn't correspond to the natural order of
the values.