Made SimpleIntegerIndexStrategy use Lexicoders
The SimpleIntegerIndexStrategy now uses Short, Integer, or Long
lexicoder to encode/decode numbers. This allows for adjusting the
size of the keys (i.e. a 16 or 32 bit integer might give a large
enough range for some applications) to save space in the row ID.

Added a convenience method to the NumberLexicoder interface for
getting a byte array from a long (dropping the most significant
bits). This is useful for the SimpleIntegerIndexStrategy.

Note: SimpleIntegerIndexStrategy can't be generic since it wouldn't
be able to change its type at runtime. Therefore, it wouldn't be
able to implement the Persistable interface.